% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMallet.R
\name{build_model_interactively}
\alias{build_model_interactively}
\title{Build a topic model from scratch based on user values from an interactive console.}
\usage{
build_model_interactively()
}
\description{
Creates directories inside a user determined base directory and names files with identifiable file name relevant to the selected parameters.  Using this function does not provide access to all of Mallet's parameters.
}
