% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMallet.R
\name{get_info}
\alias{get_info}
\title{Get information about instances.}
\usage{
get_info(
  MALLET_PATH,
  input_file = "-",
  print_instances = FALSE,
  print_infogain = 0,
  print_labels = FALSE,
  print_features = FALSE,
  print_feature_counts = FALSE,
  print_matrix = "sic"
)
}
\arguments{
\item{MALLET_PATH}{path to local mallet directory}

\item{input_file}{Read the instance list from this file; Using - indicates stdin}

\item{print_instances}{Print labels and contents for all instances. TRUE|FALSE}

\item{print_infogain}{Print top N words by information gain, sorted. Default is 0}

\item{print_labels}{Print class labels known to instance list, one per line. TRUE|FALSE}

\item{print_features}{Print the data alphabet, one feature per line. TRUE|FALSE}

\item{print_feature_counts}{Print feature names, feature counts (ie term frequency), and feature index counts (ie document frequency). TRUE|FALSE}

\item{print_matrix}{Print word/document matrix in the specified format (a|s)(b|i)(n|w|c|e), for (all vs. sparse), (binary vs. integer), (number vs. word vs. combined vs. empty) Defaut is sic}
}
\description{
A tool for printing information about instance lists of feature vectors.
}
