% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMallet.R
\name{get_top_docs}
\alias{get_top_docs}
\title{Get top documents for given topic.}
\usage{
get_top_docs(doc_topics_path, topic_index, rows = 5)
}
\arguments{
\item{doc_topics_path}{location of doc_topics output from calling train_topics}

\item{topic_index}{The index of the target topic of interest: e.g. 1:n. Note that topics are reindexed from one to be more R-like.  e.g. Mallet topic 0 is now topic 1.}

\item{rows}{The number of documents to return.}
}
\description{
use a trained topic model to infer topics for new documents
}
