% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMallet.R
\name{get_topic_keys}
\alias{get_topic_keys}
\title{Get the topic keys (i.e. top words) for a given topic.}
\usage{
get_topic_keys(topic_keys_path, topic_index)
}
\arguments{
\item{topic_keys_path}{location of keys output from calling train_topics}

\item{topic_index}{The index of the target topic of interest: e.g. 1:n. Note that topics are re-indexed from 1 to n so as to be more R-like.  e.g. Mallet topic 0 is now topic 1.}
}
\description{
use a trained topic model to infer topics for new documents
}
