% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMallet.R
\name{get_word_counts}
\alias{get_word_counts}
\title{Get the word counts per topic}
\usage{
get_word_counts(word_counts_path)
}
\arguments{
\item{word_counts_path}{location of topic word weights file output from calling train_topics}
}
\description{
Extract the word counts for each topic
}
