% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMallet.R
\name{get_word_weights}
\alias{get_word_weights}
\title{Get the topic word weights}
\usage{
get_word_weights(word_weights_path)
}
\arguments{
\item{word_topic_counts_path}{location of topic word weights file output from calling train_topics}
}
\description{
Extract the word weights file.
}
