% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMallet.R
\name{infer_topics}
\alias{infer_topics}
\title{Run MALLET infer-topics function.}
\usage{
infer_topics(
  MALLET_PATH,
  bookId,
  input_string,
  pipe_from_path,
  inferencer_path
)
}
\arguments{
\item{MALLET_PATH}{path to local mallet directory}

\item{bookId}{id of document for which topics are being inferred}

\item{input_string}{a string of words to infer on}

\item{pipe_from_path}{path to instances file created during import_dir}

\item{inferencer_path}{path to inferencer file created during train_topics}
}
\description{
use a trained topic model to infer topics for new documents
}
