% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMallet.R
\name{prune}
\alias{prune}
\title{Prune vocabulary}
\usage{
prune(
  input = "-",
  output = "-",
  prune_count = 0,
  prune_doc_freq = 0,
  min_idf = 0,
  max_idf = Inf
)
}
\arguments{
\item{input}{path to mallet.instances file}

\item{output}{path to write pruned instances file}

\item{prune_count}{Reduce features to those that occur more than N times.}

\item{prune_doc_freq}{Reduce features to those that occur in more than N contexts}

\item{min_idf}{Remove frequent features using inverse document frequency less than this value.}

\item{max_idf}{Remove rare features with inverse document frequency greater than this value.}
}
\value{
path to the resulting pruned instances file.
}
\description{
Prune removes features based on frequency or information gain.  This is an impoverished version of the core function in Mallet that allows for more options.
}
