% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMallet.R
\name{train_topics}
\alias{train_topics}
\title{Run MALLET train-topics function.}
\usage{
train_topics(
  MALLET_PATH,
  input_file,
  output_dir,
  num_topics,
  num_iterations = 1000,
  random_seed = 0,
  num_top_words = 20,
  topic_word_weights_file = NULL,
  word_topic_counts_file = NULL,
  num_threads = 1
)
}
\arguments{
\item{MALLET_PATH}{path to local mallet directory}

\item{input_file}{path to file output from calling import_dir or import_file}

\item{output_dir}{path to output directory}

\item{num_topics}{the number of topics to collect}

\item{num_iterations}{The number of iterations of Gibbs sampling.}

\item{random_seed}{useful to make process repeatable}

\item{num_top_words}{the number of top words to show in console and keys file}

\item{topic_word_weights_file}{a file path}

\item{word_topic_counts_file}{a file path}

\item{num_threads}{number of threads to use.}
}
\description{
train a topic model from Mallet data files
}
